Changing byte_size from size_t to int in heap-inl.h (making win64 build work) Review URL: http://codereview.chromium.org/1619017
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@4415 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.h
@@ -955,7 +955,7 @@ class Heap : public AllStatic {
 
   // Optimized version of memmove for blocks with pointer size aligned sizes and
   // pointer size aligned addresses.
-  static inline void MoveBlock(Object** dst, Object** src, size_t byte_size);
+  static inline void MoveBlock(Object** dst, Object** src, int byte_size);
 
   // Check new space expansion criteria and expand semispaces if it was hit.
   static void CheckNewSpaceExpansionCriteria();
